requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-suit==0.2.7
djangorestframework==3.14.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
yapf==0.32.0
